refactor : 문서 정리
