+++ docs/First_PR.md
@@ -1,0 +1,62 @@
+### 요구사항
+- 콘솔로 구현하였습니다.
+    - [x]  더하기
+    - [x]  빼기
+    - [x]  곱하기
+    - [x]  나누기
+    - [x]  우선순위(사칙연산)
+- [x]  테스트 코드 구현하기
+- [x]  계산 이력을 맵으로 데이터 저장기능 만들기
+    - 애플리케이션이 동작하는 동안 데이터베이스 외에 데이터를 저장할 수 있는 방법을 고안해보세요.
+- [x] 정규식 사용
+
+## console Package
+
+- 패키지명 그대로 사용자 쓰고 읽는 콘솔 부분을 담당하는 패키지입니다.
+- 콘솔 그대로 값 입력(InputView.class) 및 출력(OutputView.class)의 책임만 갖도록 구현하였습니다.
+
+## domain Package
+
+- 저장소와 해당 저장소를 담을 모델을 갖고 있는 패키지입니다.
+
+    ### model Package
+
+  - Calculator.class : 저장소에 담을 모델입니다.
+    - 연산식을 가집니다.
+    - 해당 연산식의 정답을 가집니다.
+  - SymbolPriority : 사칙연산을 정의한 모델입니다. 
+    - 책임
+      - 사칙연산의 우선순위를 가집니다.
+      - 사칙연산의 역할을 가집니다.
+
+  ### repository Package
+    
+  - CalculationRepository.class : 모델(Calculator.class)을 담는 저장소입니다.
+  - IdGenerator.class : 저장소의 기본키 역할을 하고 있습니다.
+    - 기본키의 책임만을 가지고 있고 싱글톤으로 구현하였습니다.
+    - ID 값을 가지고 올 때마다 자동으로 증가되기 때문에, 무결성을 보장합니다.
+    
+### error Package
+
+- 예러에 대한 응답을 담는 패키지입니다.
+- ResponseErrorFormat : 예외 발생 시, 응답을 공통화하기 위한 클래스입니다.
+  - 필요한 예외들에 대해서 응답 메시지를 양식화하였습니다.
+  - 따로 예외 클래스를 만들어 처리하지 않은 이유는 Over Spec 이라 생각하였고 이 양식 하나만으로 충분히 표현할 수 있을 것이라 생각했습니다.
+
+### service
+
+- 각 요구 기능을 실질적으로 처리하는 패키지입니다.
+- 즉, CalculationService.class 는 사용자의 요구사항을 실질적으로 처리하는 책임을 가집니다.
+- 저는 인터페이스가 관련 없는 객체 간 자신에게 맞게 구현하는데 목적을 두어 유연하고 느슨한 연결을 위해 사용하는 것으로 알고 있습니다. 때문에, 현재 각 서비스의 로직들은 하나의 계산기를 위해 하나의 목적만을 가지고 있어서, 인터페이스를 사용하지 않았습니다.
+- 실제로, 요구 기능을 위해 사용되는 메소드는 최대한 책처럼 읽을 수 있도록 구현하였습니다.
+  - 조회 기능 : void getCalculationsAll() { ... }
+  - 계산 기능 : void calculate(String operation) { ... }
+  - 실질적으로 연산이 이루어지는 메소드는 processOperatorsAndNumbers(String operation) { ... } 입니다.
+
+### CalculatorApplication.class
+
+- 해당 어플리케이션을 실행하는 곳입니다.
+- 사용자에게 출력 및 입력을 요구하는 뷰를 가져오는 책임을 가집니다.
+- 사용자가 원하는 기능을 호출하는 책임을 가집니다.
+- 사용자가 입력한 값을 가지고 원하는 서비스를 호출 했을 때 해당 서비스에서 예외가 발생하면 catch 하는 책임을 가집니다.
+- 사용자가 입력한 값을 가지고 해당 호출한 서비스가 정상적으로 동작하면 출력하는 책임을 가집니다.
