+++ src/test/java/org/domain/repository/IdGeneratorTest.java
@@ -1,6 +1,5 @@
 package org.domain.repository;
 
-import org.domain.repository.IdGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
